eventloop work and general cleanup

+++ include/time.h
@@ -3,8 +3,8 @@
 
 
 namespace time {
-	unsigned long now_us(void);
-	unsigned long now_ms(void);
+	unsigned long long now_us(void);
+	unsigned long long now_ms(void);
 
 	// called by the timekeeper thread on tick
 	void timekeep();
